[py] Fix installing most of the data from source distributions
Change `namespaces` to true to fix installing
`selenium.webdriver.common.devtools.*` packages, and fix `package-data`
to fix installing `.js` and `.json` data files.  The latter would
specify a non-existing `selenium_package` package name -- instead,
specify `*` to apply it to all Python packages found.

With these changes, the result of installing the source distribution is
almost identical to the result of installing the wheel.  One remaining
problem is that `selenium-manager` is not included -- my guess is that
there is no logic copying the built Rust executable into the expected
location.

Another problem is that these changes do not work when installing from
a VCS checkout -- since all the relevant data files are explicitly
excluded by `.gitignore` and that takes precedence over `data-files`.
However, this was broken before, so it should not really matter.

Fixes #15125