service: only shutdown if process not terminated
There are scenarios where a stop() is called on a Service object
multiple times. This also happens when explicitly quitting a Webdriver using
driver.quit() and afterwards having the garbage collector destroy the service
object, calling stop() another time even though the service process
has already terminated.

The check inside the stop() call only ensured that the process variable
is not None, but ignored the fact that the process might already have
terminated. Therefor an additional check is introduced to only send
the remote shutdown command if the service process has not ended.

Fixes #15182

Signed-off-by: Sandro Pischinger <mail@sandropischinger.de>